+++ SRC/Private/Comparers/ServiceIDComparer.cs
@@ -8,7 +8,7 @@ namespace Solti.Utils.DI.Internals
     internal sealed class ServiceIdComparer : ComparerBase<ServiceIdComparer, IServiceId>
     {
         public override int GetHashCode(IServiceId obj) =>
-#if NETSTANDARD1_6 || NETSTANDARD2_0
+#if NETSTANDARD2_0
             new { obj.Interface, obj.Name }.GetHashCode()
 #else
             System.HashCode.Combine(obj.Interface, obj.Name)