+++ SRC/Public/Patterns/ServiceContainer.cs
@@ -9,9 +9,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-#if NETSTANDARD1_6
-using System.Reflection;
-#endif
 using System.Threading;
 using System.Threading.Tasks;