+++ history.md
@@ -98,4 +98,5 @@
 - 2.0.0-preview1:
   - *breaking:* removed `ServiceReference.DisposeSupressed` property
   - *breaking:* `ServiceContainer.GetEnumerator()` no longer clones the container (enumerating is still thread safe)
+  - *breaking:* dropped `netstandard1.6` support
   - *added:* `IAsyncDisposable` support